Fix exponential content observer registration
On each call to ContentObserver.onChange(), updateUi() re-registers a
ContentObserver. On the first onChange, a second observer is
registered and for the next change, both observers call updateUi()
registering two more, and so on. After n onChanges there are 2^n
observers.

Remove the duplicate registrations and register a ContentObserver only
if the station changes.

FEIJ-1373

Change-Id: I07d0bd968165dfda5601f72fd20641bb9d62a5f1
Signed-off-by: Louis Popi <theh2o64@gmail.com>